+++ _src/interfaces/streamServiceExecutor.ts
@@ -1,6 +1,0 @@
-import { Observable } from 'rxjs';
-import { ExecutionContext } from './executionContext';
-
-export interface StreamServiceExecutor<TRequest, TResponse> {
-	stream(request: TRequest, context?: ExecutionContext): Observable<TResponse>;
-}