+++ _src/managers/providers/browser.ts
@@ -1,4 +1,3 @@
-import { Observable, Observer } from 'rxjs';
 import { ServiceExecutor } from '../../interfaces/serviceExecutor';
 import { ExecutionContext } from '../../interfaces/executionContext';
 import { NaniumRepository } from '../../interfaces/serviceRepository';
@@ -103,7 +102,6 @@ export class NaniumProviderBrowser implements ServiceProviderManager {
 				request = NaniumObject.create(request, requestConstructor);
 			}
 
-
 			// interceptors
 			if (this.config.requestInterceptors?.length) {
 				let result: any;
@@ -123,46 +121,6 @@ export class NaniumProviderBrowser implements ServiceProviderManager {
 		} catch (e) {
 			return await this.config.handleError(e, serviceName, request, context);
 		}
-	}
-
-	stream(serviceName: string, request: any, context?: ExecutionContext): Observable<any> {// validation
-		context = context || {};
-
-		if (this.repository === undefined) {
-			return this.createErrorObservable(new Error('nanium service repository is not initialized'));
-		}
-		if (!Object.prototype.hasOwnProperty.call(this.repository, serviceName)) {
-			return this.createErrorObservable(new Error('unknown service ' + serviceName));
-		}
-		const requestConstructor: any = this.repository[serviceName].Request;
-		const realRequest: any = new requestConstructor();
-		Object.assign(realRequest, request);
-		if (context && context.scope === 'public') { // private is the default, all adaptors have to set the scope explicitly
-			if (!requestConstructor.scope || requestConstructor.scope !== 'public') {
-				return this.createErrorObservable(new Error('unauthorized'));
-			}
-		}
-
-		return new Observable<any>((observer: Observer<any>): void => {
-			new this.repository[serviceName].Executor().stream(realRequest, context).subscribe({
-				next: (value: any): void => {
-					observer.next(value);
-				},
-				error: (e: any): void => {
-					this.config.handleError(e, serviceName, realRequest, context).then();
-					observer.error(e);
-				},
-				complete: (): void => {
-					observer.complete();
-				}
-			});
-		});
-	}
-
-	private createErrorObservable(e: any): Observable<any> {
-		return new Observable((observer: Observer<any>): void => {
-			observer.error(e);
-		});
 	}
 
 	async emit(eventName: string, event: any, executionContext: ExecutionContext): Promise<void> {